# feat: optimize LZMA processing with enhanced multiReader implementation for io.ByteReader