This pull request aims to optimize the LZMA2 processing in the `github.com/bodgit/sevenzip` project by enhancing the usage of `github.com/ulikunitz/xz` library for io.ByteReader. 

The current implementation doesn't fully leverage the optimizations provided by `github.com/ulikunitz/xz` due to its usage of io.MultiReader. As a result, the proposed change modifies the code to return a custom implementation of multiReader based on io.ReadCloser when io.ByteReader is implemented.

The optimized multiReader implementation exhibits significant performance improvements, especially when handling large 7z archives. While the impact might be limited for smaller inputs, it can provide an efficiency boost of nearly 10% for larger files.

before:
```
goos: linux
goarch: amd64
pkg: github.com/bodgit/sevenzip
cpu: Intel(R) Core(TM) i7-8650U CPU @ 1.90GHz
BenchmarkLargeLZMA2-8   	       1	323613023997 ns/op
BenchmarkLZMA2-8        	     816	   1648719 ns/o
```

after:
```
goos: linux
goarch: amd64
pkg: github.com/bodgit/sevenzip
cpu: Intel(R) Core(TM) i7-8650U CPU @ 1.90GHz
BenchmarkLargeLZMA2-8   	       1	307210122907 ns/op
BenchmarkLZMA2-8        	     926	   1512918 ns/op
```

I kindly request your review and feedback on this pull request. Thank you for considering this contribution.