@bodgit 2023-06-19
I'm slightly confused as your PR and benchmark performance references LZMA2, but the reader you've changed is only used for LZMA(1) streams; there's a separate reader used for LZMA2.