@orisano 2023-06-19
@bodgit thanks for reply.
Sorry for your confusion.
The folder 0 of testdata/lzma2.7z is compressed with lzma and the lzma reader is called instead of lzma2. 
The 7z I am working with now has folder 0 as lzma as testdata/lzma2.7z, and its large size affects performance.