@bodgit 2023-07-03
I've merged this PR and released it as v1.4.3. If/when the xz package is updated and contains your fix, I'll create a further release.

Thanks for the contribution!